Javascript changes
Added a Set to keep track of clicked zones. Updated the mousePressed() function to add the clicked zone to the set and check if all zones have been clicked. Added functions for resetting the game and continuing. Added a div for the end popup to display when all zones are clicked. Added buttons for starting over and continuing. Thank you Chat GPT